Consolidate Vega sandbox constants
Move VEGA_SANDBOX_ROUTE_PATH into common/constants.ts and update imports, removing the separate sandbox_constants module.
