+++ src/platform/plugins/private/vis_types/vega/public/vega_view/vega_sandbox_frame_host.ts
@@ -9,7 +9,7 @@
 
 import type { HttpStart } from '@kbn/core/public';
 import type { VegaSandboxMessage } from '@kbn/vega-sandbox';
-import { VEGA_SANDBOX_ROUTE_PATH } from '../../common/sandbox_constants';
+import { VEGA_SANDBOX_ROUTE_PATH } from '../../common/constants';
 
 const SANDBOX_ATTRIBUTE = 'allow-scripts';
 const DISALLOWED_SANDBOX_TOKENS = ['allow-same-origin', 'allow-top-navigation', 'allow-popups'];
