+++ src/platform/plugins/private/vis_types/vega/server/sandbox_route.test.ts
@@ -12,7 +12,7 @@ import {
   getImgSrcSourcesFromPolicy,
   registerSandboxRoute,
 } from './sandbox_route';
-import { VEGA_SANDBOX_ROUTE_PATH } from '../common/sandbox_constants';
+import { VEGA_SANDBOX_ROUTE_PATH } from '../common/constants';
 
 describe('Vega sandbox route CSP helpers', () => {
   test('maps broad allow-all policy to wildcard image source when there are no deny rules', () => {
